Fix for image download, remove wrong jpgs easier

+++ src/remove_wrong_images.py
@@ -1,17 +1,18 @@
 import os
 from PIL import Image
 
+# Definiowanie ścieżek
+base_dir = 'data/pre_images'
 
-image_dirs = ['data/train', 'data/validation']
-
-for image_dir in image_dirs:
-    for folder in os.listdir(image_dir):
-        folder_path = os.path.join(image_dir, folder)
-        for image_name in os.listdir(folder_path):
-            image_path = os.path.join(folder_path, image_name)
+# Przechodzenie przez każdy folder (gatunek) w folderze bazowym
+for folder in os.listdir(base_dir):
+    current_dir = os.path.join(base_dir, folder)
+    if os.path.isdir(current_dir):
+        for image_name in os.listdir(current_dir):
+            image_path = os.path.join(current_dir, image_name)
             try:
                 img = Image.open(image_path)
-                img.verify()  # Potwierdź, plik jest obrazem
+                img.verify()  # Potwierdz, ze plik jest obrazem
             except (IOError, SyntaxError):
                 print('Uszkodzony plik:', image_path)
                 os.remove(image_path)
